# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salono Backend API",
        "description": "Documentation de l'API Salono Backend",
        "contact": {
            "email": "support@salono.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/v1/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Réinitialiser le mot de passe",
                "description": "Réinitialise le mot de passe avec le token reçu après vérification du code",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "Token et nouveau mot de passe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "abc123token"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "NouveauMotDePasse123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NouveauMotDePasse123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre mot de passe a été réinitialisé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides ou token invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Le token de réinitialisation est invalide ou a expiré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Impossible de réinitialiser le mot de passe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token JWT obtenu lors de la connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}